wip: moved into separate config structs. builds will use structs from build package
