+++ backend/pkg/vcli/runner.go
@@ -3,8 +3,6 @@ package vcli
 import (
 	"fmt"
 	"sync"
-
-	"github.com/velocity-ci/velocity/backend/pkg/velocity/task"
 )
 
 type runner struct {
@@ -20,45 +18,45 @@ func newRunner(wg *sync.WaitGroup) *runner {
 }
 
 func (r *runner) Run(taskName string) {
-	r.run = true
-	defer r.wg.Done()
-	defer func() { r.run = false }()
-	tasks, _ := task.GetTasksFromCurrentDir()
+	// r.run = true
+	// defer r.wg.Done()
+	// defer func() { r.run = false }()
+	// tasks, _ := task.GetTasksFromCurrentDir()
 
-	var t *task.Task
-	// find Task requested
-	for _, tsk := range tasks {
-		if tsk.Name == taskName {
-			t = tsk
-			break
-		}
-	}
+	// var t *task.Task
+	// // find Task requested
+	// for _, tsk := range tasks {
+	// 	if tsk.Name == taskName {
+	// 		t = tsk
+	// 		break
+	// 	}
+	// }
 
-	if t == nil {
-		fmt.Printf("Task %s not found in:\n%v\n", taskName, tasks)
-		return
-	}
-	fmt.Printf("Running task: %s\n", t.Name)
+	// if t == nil {
+	// 	fmt.Printf("Task %s not found in:\n%v\n", taskName, tasks)
+	// 	return
+	// }
+	// fmt.Printf("Running task: %s\n", t.Name)
 
-	emitter := NewEmitter()
+	// emitter := NewEmitter()
 
-	t.Steps = append([]task.Step{task.NewSetup()}, t.Steps...)
+	// t.Steps = append([]task.Step{task.NewSetup()}, t.Steps...)
 
-	// Run each step unless they fail (optional)
-	for i, step := range t.Steps {
-		if !r.run {
-			return
-		}
-		if step.GetType() == "setup" {
-			step.(*task.Setup).Init(&ParameterResolver{}, nil, "")
-		}
-		emitter.SetStepNumber(uint64(i))
-		err := step.Execute(emitter, t)
-		if err != nil {
-			fmt.Printf("encountered error: %s", err)
-			return
-		}
-	}
+	// // Run each step unless they fail (optional)
+	// for i, step := range t.Steps {
+	// 	if !r.run {
+	// 		return
+	// 	}
+	// 	if step.GetType() == "setup" {
+	// 		step.(*task.Setup).Init(&ParameterResolver{}, nil, "")
+	// 	}
+	// 	emitter.SetStepNumber(uint64(i))
+	// 	err := step.Execute(emitter, t)
+	// 	if err != nil {
+	// 		fmt.Printf("encountered error: %s", err)
+	// 		return
+	// 	}
+	// }
 }
 
 func (r *runner) Stop() {
